CV-tailor-master/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
openai==2.14.0
python-dotenv==1.2.1
python-docx==1.2.0
pydantic==2.12.5
pydantic[email]==2.12.5
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
email-validator==2.1.0
